Create README.md
As the Promotion Rule would be changing often, designing it as a Functional Interface provides it the ability to change with the demands of the market, hence.  Without a real database, a lot of static initializations are adopted to provide data. I used Builder as Products tend to have many fields.  even though this Product has just
 a few fields.. Initially I was quite sure if you expect me to build as a web application, but after reading the test closely. I decided to write it as a  jar file Java Application with no Web features. I thought it was the design and functionality of the Promotion Rules that is more relevant.

It runs without compilation errors and produces the required results as described in the test.   I embedded testing while developing though execution results as this were class level testing with no integration with infrastructure. However,  I tested them vigorously with edge data to make sure it is extensively tested.